Major code refactor
The bot was completely reorganized in how it handles messages, databases, and overall methods. To begin, all of the important objects needed for a command was placed inside an object called ctx, which now stores the message, channel, guild, args, user, client, and utils, which is another addition that provides methods for getting random numbers, creating embeds, escaping characters and so forth. A class called Account was also made, which stores and loads the database values for a specific account into one object, and can be used to synchronize with the database once all of the values have been edited as needed. I also cleaned up all of the commands overall, and the name 'help' in  is now . The only downside is that I haven't been able to test the new layout, so there is a possibility of something being incorrectly written in here.