Cleanup and Notes
Cleaned up more of the help command, and connected the utils.setAuthor() method made earlier to this. New ideas include:
Creating an 'info.args' array, which stores information for each argument that is needed for the command. This way, you can try to parse the arguments for a command BEFORE it runs, which can handle all of the errors at that point.